feat: update main component
